requirements.txt
fastapi
uvicorn
python-multipart

ffmpeg-python
pydub
librosa
soundfile

openai-whisper
torch

nltk
spacy
scikit-learn

requests
python-dotenv
tqdm
rich
email-validator